python-backend/requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==1.10.12
bcrypt==4.1.3
jwt==1.3.1
scikit-learn==1.3.2
spacy>=3.0.0
sentence-transformers>=2.0.0
newsapi-python>=0.2.6
langdetect>=1.0.0
python-multipart>=0.0.5